Compilers_Principles/experiment9/requirements.txt
ply~=3.11
pytest~=7.4.3
fpdf2~=2.7.6